extract strategies from passport middleware
