+++ server/middlewares/passport.js
@@ -3,34 +3,8 @@ const __ = CONFIG.universalPath
 const _ = __.require('lib', 'utils')
 const session = require('express-session')
 const passport = require('passport')
-, OAuth2Strategy = require('passport-oauth').OAuth2Strategy;
 const User = __.require('controllers', '/users/lib/users')
-const breq = require('bluereq')
-
-const GitlabOAuth2Strategy = new OAuth2Strategy({
-  authorizationURL: CONFIG.get('auth.gitlab.authorizationURL'),
-  tokenURL: CONFIG.get('auth.gitlab.tokenURL'),
-  clientID: CONFIG.get('auth.gitlab.clientID'),
-  clientSecret: CONFIG.get('auth.gitlab.clientSecret'),
-  callbackURL: CONFIG.get('auth.gitlab.callbackURL'),
-}, function(accessToken, refreshToken, profile, done) {
-  // smell: `profile` should be filled with user info, instead requesting user info through accessToken endpoint
-  accessTokenEndpoint = "https://lab.allmende.io/api/v3/user?access_token="
-  breq.get( accessTokenEndpoint + accessToken )
-  .then(function (res) {
-    const userInfo = {
-      contact: {
-        name: res.body.username,
-        email: res.body.email,
-      },
-      auth: {
-        provider: res.body.web_url
-      }
-    }
-    const user = User.findOrCreateUser(userInfo)
-    done(null, user)
-  })
-})
+const gitlabStrategy = require('./strategies/gitlab')
 
 passport.serializeUser(function(user, done) {
   _.success(id = user._id, 'serializeUser')
@@ -48,7 +22,7 @@ passport.deserializeUser(function(id, done) {
   })
 })
 
-passport.use('gitlab', GitlabOAuth2Strategy);
+passport.use('gitlab', gitlabStrategy.auth)
 
 const authenticate = passport.authenticate('gitlab', { failureRedirect: '/', successReturnToOrRedirect: '/' })
